refactor: 앨범 상세 Model 분리 #149
